SM: App storage export created - added create/delete task functions.
